+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Introduce `HashLike` module and make STAC Object classes include it.
 - Fix a bug that `Catalog#all_items` returns only direct children's items.
 - Improve documentation.
+- Add methods `#add_child` and `#add_item` to `STAC::Catalog`
 
 ## [0.2.0] - 2022-11-02